GlobalRadioPlayer/src/requirement.txt
requests>=2.32.0
python-vlc>=3.0.18121
charset-normalizer>=3.3.0
